requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
analytics-python==1.4.post1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
backoff==1.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
black==19.10b0
boto3==1.26.101
botocore==1.29.101
braceexpand==0.1.7
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.2
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
comtypes==1.2.0
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
Cython==3.0.2
debugpy==1.6.6
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.19
editdistance==0.6.2
einops==0.6.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.3
fastapi==0.95.0
fasttext==0.9.2
ffmpy==0.3.0
filelock==3.10.7
Flask==2.2.3
Flask-RESTful==0.3.9
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.0
google-auth-oauthlib==0.4.6
gradio==3.4
grpcio==1.53.0
h11==0.12.0
h5py==3.8.0
httpcore==0.15.0
httpx==0.23.3
huggingface-hub==0.13.3
hydra-core==1.2.0
idna==3.4
ijson==3.2.0.post0
imagesize==1.4.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflect==6.0.2
iniconfig==2.0.0
ipadic==1.0.0
ipykernel==6.22.0
ipython==8.11.0
ipywidgets==8.0.6
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
jellyfish==1.0.0
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.2
jmespath==1.0.1
joblib==1.2.0
jupyter-client==8.1.0
jupyter-core==5.3.0
jupyterlab-widgets==3.0.7
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.4.4
kornia==0.6.11
language-tool-python==2.7.1
latexcodec==2.0.1
lazy-loader==0.2
Levenshtein==0.20.9
librosa==0.10.0.post2
lightning-utilities==0.8.0
linkify-it-py==2.0.0
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mecab-python3==1.0.5
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nemo-toolkit==1.16.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.13.1
OpenCC==1.1.1
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pangu==4.0.6.1
parameterized==0.9.0
paramiko==3.1.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.4.0
plac==1.3.5
platformdirs==3.2.0
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
progress==1.6
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.0
pyannote.metrics==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pybind11==2.10.4
pybtex==0.24.0
pybtex-docutils==1.0.2
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydub==0.25.1
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pypinyin==0.48.0
pypinyin-dict==0.5.0
pypiwin32==223
PySocks==1.7.1
pytest==7.2.2
pytest-runner==6.0.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.9.0
pyttsx3==2.90
pytz==2023.3
pywin32==306
PyYAML==5.4.1
pyzmq==25.0.2
rapidfuzz==2.13.7
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.18.0
setproctitle==1.3.2
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
sox==1.4.1
soxr==0.3.4
sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
starlette==0.26.1
strsimpy==0.2.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
text-unidecode==1.3
textblob==0.17.1
textdistance==4.5.0
texterrors==0.4.4
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.3
typed-ast==1.5.4
typer==0.7.0
typing-extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
wandb==0.14.0
wcwidth==0.2.6
webdataset==0.1.62
websockets==10.4
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.7
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.8.2
youtokentome==1.0.6
zipp==3.15.0
